wrap constructor in httpclient class
